Lost Data Recovery...
Recovering the lost data. Rather than redoing all of the SQL code, just finishing up the New Client Form's UI, then starting to do the code behind the scenes.
