+++ EstateMastersInstitute/ClientLookupForm.cs
@@ -16,5 +16,11 @@ namespace EstateMastersInstitute
         {
             InitializeComponent();
         }
+
+        private void closeButton_Click(object sender, EventArgs e)
+        {
+            this.Close();
+
+        }
     }
 }
