--- EstateMastersInstitute/NewClientForm.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace EstateMastersInstitute
-{
-    public partial class EditClientForm : Form
-    {
-        public EditClientForm()
-        {
-            InitializeComponent();
-        }
-
-        private void EditClientForm_Load(object sender, EventArgs e)
-        {
-
-        }
-    }
-}
